Replace PetSwiper/Unswiper/Rejector with PetInteractor
