+++ src/main/java/server/use_cases/pet_interactor/PetInteractorRequestModel.java
@@ -1,13 +1,13 @@
-package server.use_cases.pet_unswiper;
+package server.use_cases.pet_interactor;
 
 import server.use_cases.AuthRequestModel;
 
-public class PetUnswiperRequestModel extends AuthRequestModel {
+public class PetRejectorRequestModel extends AuthRequestModel {
 
     private String pet1Id;
     private String pet2Id;
 
-    public PetUnswiperRequestModel(String headerUserId, String pet1Id, String pet2Id) {
+    public PetRejectorRequestModel(String headerUserId, String pet1Id, String pet2Id) {
         super(headerUserId);
         this.pet1Id = pet1Id;
         this.pet2Id = pet2Id;
