--- src/main/java/server/use_cases/pet_swiper/PetSwiperInputBoundary.java
@@ -1,9 +1,0 @@
-package server.use_cases.pet_swiper;
-
-import server.use_cases.ResponseModel;
-
-public interface PetSwiperInputBoundary {
-
-    public ResponseModel swipe(PetSwiperRequestModel request);
-
-}
