--- src/main/java/server/use_cases/pet_unswiper/PetUnswiperInputBoundary.java
@@ -1,9 +1,0 @@
-package server.use_cases.pet_unswiper;
-
-import server.use_cases.ResponseModel;
-
-public interface PetUnswiperInputBoundary {
-
-    public ResponseModel unswipePets(PetUnswiperRequestModel request);
-
-}
